+++ Next Wave/Views/NearestStationTileView.swift
@@ -138,7 +138,7 @@ struct NearestStationTileView: View {
                                     .font(.system(size: 11))
                                     .foregroundColor(Color("text-color"))
                                 
-                                Text(String(format: "%.0f°", waterTemp))
+                                Text(String(format: "%.1f°", waterTemp))
                                     .font(.system(size: 11))
                                     .foregroundColor(Color("text-color"))